refactor: view 폴더 분리
